refactor(llm,config,scheduler,core): code quality improvements
Split monolithic source files into focused submodules:
- zeph-vault: extract age/arc/env/mock backends from lib.rs (1302 → ~150 lines)
- zeph-sanitizer: extract ContentSanitizer into sanitizer.rs (2313 → ~150 lines)
- zeph-config: split 5080-line migrate.rs into migrate/mod.rs + migrate/steps.rs

Replace wildcard `use super::*` and `use steps::*` in migrate split with
explicit imports to satisfy clippy::wildcard_imports.

Standardize deferred-refactor comment markers across 41 files to a
searchable `TODO(review):` prefix.

Fix error message casing in zeph-skills and zeph-index error enums.
