+++ crates/zeph-context/src/assembler.rs
@@ -87,7 +87,7 @@ impl ContextAssembler {
     /// # Errors
     ///
     /// Propagates errors from any async fetch operation.
-    #[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — deferred to a future structural refactor
+    #[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — TODO(review): file a tracking issue for this decomposition
     pub async fn gather(input: &ContextAssemblyInput<'_>) -> Result<PreparedContext, ContextError> {
         type CtxFuture<'a> =
             Pin<Box<dyn Future<Output = Result<ContextSlot, ContextError>> + Send + 'a>>;
