+++ crates/zeph-core/src/agent/tool_execution/tests.rs
@@ -2897,7 +2897,7 @@ async fn self_reflection_single_tool_failure_produces_one_tool_result() {
 // Second tool fails → reflection fires → early return must append ToolResult for 2nd (is_error)
 // and a synthetic [skipped] ToolResult for the 3rd. Total: 3 ToolResults for 3 ToolUses.
 #[tokio::test]
-#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — deferred to a future structural refactor
+#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — TODO(review): file a tracking issue for this decomposition
 async fn self_reflection_middle_tool_failure_no_orphans() {
     use std::sync::{Arc, Mutex};
 
@@ -3566,7 +3566,7 @@ async fn transient_error_on_non_retryable_executor_is_not_retried() {
 // tool[2] is non-retryable-transient-always-fail. Verifies all three complete with
 // the correct outcome and the retry fires only for tool[1].
 #[tokio::test]
-#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — deferred to a future structural refactor
+#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — TODO(review): file a tracking issue for this decomposition
 async fn mixed_retryable_and_non_retryable_batch() {
     use super::super::agent_tests::{MockChannel, create_test_registry, mock_provider};
     use std::sync::atomic::{AtomicUsize, Ordering};
@@ -4290,7 +4290,7 @@ async fn sanitize_tool_output_memory_save_still_uses_tool_result() {
 // After the fix, both ToolResults must be present in a single User message that immediately
 // follows the Assistant{ToolUse} message, with no interleaved messages in between.
 #[tokio::test]
-#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — deferred to a future structural refactor
+#[allow(clippy::too_many_lines)] // long function; decomposition would require extracting state into additional structs — TODO(review): file a tracking issue for this decomposition
 async fn test_parallel_tool_calls_permanent_error_emits_tool_result() {
     use std::sync::Arc;
     use std::sync::atomic::{AtomicUsize, Ordering};
